refactor: rename DashboardCardCell to DashboardPostsCardCell
